move all files data fillers
