+++ src/ReportingBundle/Resources/config/services.yml
@@ -1,7 +1,7 @@
 services:
 
     ############################### Data Fillers ###############################
-    
+
     reporting.data_fillers.reference:
         class: ReportingBundle\Utils\DataFillers\DataFillersIndicator
         arguments: ['@doctrine.orm.entity_manager']
@@ -13,17 +13,17 @@ services:
         public: true
 
     reporting.data_fillers.country:
-        class: ReportingBundle\Utils\DataFillers\DataFillersCountry
+        class: ReportingBundle\Utils\DataFillers\Country\DataFillersCountry
         arguments: ['@doctrine.orm.entity_manager']
         public: true
 
     reporting.data_fillers.project:
-        class: ReportingBundle\Utils\DataFillers\DataFillersProject
+        class: ReportingBundle\Utils\DataFillers\Project\DataFillersProject
         arguments: ['@doctrine.orm.entity_manager']
         public: true
 
     reporting.data_fillers.distribution:
-        class: ReportingBundle\Utils\DataFillers\DataFillersDistribution
+        class: ReportingBundle\Utils\DataFillers\Distribution\DataFillersDistribution
         arguments: ['@doctrine.orm.entity_manager']
         public: true
